% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_objects.R
\name{LicenseAssignment}
\alias{LicenseAssignment}
\title{Enterprise License Manager API Objects 
Licensing API to view and manage license for your domain.}
\usage{
LicenseAssignment(etags = NULL, productId = NULL, selfLink = NULL,
  skuId = NULL, userId = NULL)
}
\arguments{
\item{etags}{ETag of the resource}

\item{productId}{Name of the product}

\item{selfLink}{Link to this page}

\item{skuId}{Name of the sku of the product}

\item{userId}{Email id of the user}
}
\value{
LicenseAssignment object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:00:39
filename: /Users/mark/dev/R/autoGoogleAPI/googlelicensingv1.auto/R/licensing_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
LicenseAssignment Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for LiscenseAssignment Resource
}
\seealso{
Other LicenseAssignment functions: \code{\link{licenseAssignments.patch}},
  \code{\link{licenseAssignments.update}}
}
