% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_objects.R
\name{LicenseAssignmentInsert}
\alias{LicenseAssignmentInsert}
\title{LicenseAssignmentInsert Object}
\usage{
LicenseAssignmentInsert(userId = NULL)
}
\arguments{
\item{userId}{Email id of the user}
}
\value{
LicenseAssignmentInsert object
}
\description{
LicenseAssignmentInsert Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for LicenseAssignment Insert request
}
\seealso{
Other LicenseAssignmentInsert functions: \code{\link{licenseAssignments.insert}}
}

