% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_objects.R
\name{LicenseAssignmentList}
\alias{LicenseAssignmentList}
\title{LicenseAssignmentList Object}
\usage{
LicenseAssignmentList(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{The LicenseAssignments in this page of results}

\item{nextPageToken}{The continuation token, used to page through large result sets}
}
\value{
LicenseAssignmentList object
}
\description{
LicenseAssignmentList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
LicesnseAssignment List for a given product/sku for a customer.
}

