% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_functions.R
\name{licenseAssignments.delete}
\alias{licenseAssignments.delete}
\title{Revoke License.}
\usage{
licenseAssignments.delete(productId, skuId, userId)
}
\arguments{
\item{productId}{Name for product}

\item{skuId}{Name for sku}

\item{userId}{email id or unique Id of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.licensing
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.licensing)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/licensing/}{Google Documentation}
}

