% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_functions.R
\name{licenseAssignments.insert}
\alias{licenseAssignments.insert}
\title{Assign License.}
\usage{
licenseAssignments.insert(LicenseAssignmentInsert, productId, skuId)
}
\arguments{
\item{LicenseAssignmentInsert}{The \link{LicenseAssignmentInsert} object to pass to this method}

\item{productId}{Name for product}

\item{skuId}{Name for sku}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.licensing
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.licensing)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/licensing/}{Google Documentation}

Other LicenseAssignmentInsert functions: \code{\link{LicenseAssignmentInsert}}
}
