% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_functions.R
\name{licenseAssignments.listForProductAndSku}
\alias{licenseAssignments.listForProductAndSku}
\title{List license assignments for given product and sku of the customer.}
\usage{
licenseAssignments.listForProductAndSku(productId, skuId, customerId,
  maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{productId}{Name for product}

\item{skuId}{Name for sku}

\item{customerId}{CustomerId represents the customer for whom licenseassignments are queried}

\item{maxResults}{Maximum number of campaigns to return at one time}

\item{pageToken}{Token to fetch the next page}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.licensing
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.licensing)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/licensing/}{Google Documentation}
}
