% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing_functions.R
\docType{package}
\name{licensing_googleAuthR}
\alias{licensing_googleAuthR}
\alias{licensing_googleAuthR-package}
\title{Enterprise License Manager API
Licensing API to view and manage license for your domain.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:00:39
filename: /Users/mark/dev/R/autoGoogleAPI/googlelicensingv1.auto/R/licensing_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/apps.licensing
}
}
